Add modular systems, fix framebuffer ratio, rm camera classes
Significant progress was made to allow the developer to create modular systems inside the engine, now she/he can just add whatever system it wants the moment it creates a scene and each will have its own systems. For now, the world scene and the ui scene have been set.
- fix framebuffer issue where the aspect ratio was wrong
- camera classes have been simplified to remove excessive and confusing classes which should have been encapsulated from the beginning
- ECS systems now follow a framework where they can only be initialized if the right components are correct

+++ Pain/platform/OpenGL/ShaderBackend.cpp
@@ -143,7 +143,7 @@ void uploadUniformMat4(int loc, const glm::mat4 &m)
 }
 void uploadUniformIntArray(int loc, int *values, uint32_t count)
 {
-  glUniform1iv(loc, count, values);
+  glUniform1iv(loc, static_cast<int32_t>(count), values);
 }
 
 } // namespace pain::backend
